Restore FallbackResource
It is used to load an URL when not clicking links from the home page
(e.g. when loading a page from a link or bookmarks).  The rest of
navigation was unaffected because the front UI change the address in the
address bar on each clic.

For it to work, PassengerHighPerformance must be turned off.

The current version of the apache module does not handle this properly
ATM, but this only affects users enabling PassengerHighPerformance.