Pass configruation file name via DJANGO_SETTINGS_MODULE
Previously, `settings/__init__.py` was looking for `settings/local.py`
before loading it, and defaulting to `settings/common.py` when not
found.

That behavior has been changed in Taiga 6.0.0, and now Taiga expects
that the name of the configuration file to take into account is passed
through the environment variable `DJANGO_SETTINGS_MODULE`.