Fix deployment with PostgreSQL 15
We previously did not change the database owner for PostgreSQL,
resulting in the database being owned by the `postgres` user.

PostgreSQL 15 revoked the CREATE permission from all users except a
database owner from the public schema.  As a consequence, migration that
are expected to create objects are not allowed to do so with this
version and newer of PostgreSQL.

Explicitly make the database user the owner of it, to allow migrations.